Removing dependency on brewer2mpl from color.py and adjusting choroBrewer.py
